+++ gaitmap_mad/gaitmap_mad/stride_segmentation/hmm/_hmm_stride_segmentation.py
@@ -226,11 +226,11 @@ class HmmStrideSegmentation(BaseStrideSegmentation, Generic[BaseSegmentationHmmT
         # find rising edge of stride start state sequence
         # +1 required as diff returns one element less than the input
         matches_starts = (
-            np.argwhere(np.diff((hidden_states_predicted == stride_start_state).astype(int)) > 0).flatten() + 1
+            np.argwhere(np.diff((hidden_states_predicted == stride_start_state).astype("int32")) > 0).flatten() + 1
         )
 
         # find falling edge of stride end state sequence
-        matches_ends = np.argwhere(np.diff((hidden_states_predicted == stride_end_state).astype(int)) < 0).flatten() + 1
+        matches_ends = np.argwhere(np.diff((hidden_states_predicted == stride_end_state).astype("int32")) < 0).flatten() + 1
 
         # Special case, when the last state is a stride end state
         if hidden_states_predicted[-1] == stride_end_state:
